First commit
snake.cpp
=========
The base algorithm. This implementation is based
on the following blog post by Cris Luengo.
https://www.crisluengo.net/archives/217

snake.h
=======
Main interface to access the main algorithm.

main.c
======
Main entry point to the rest of the experiment.
This might get split into several files later.

proto
=====
Same implementation in Python3 as it is in snake.cpp
but with visuals. Use this to see how the core algorithm works.

res
===
Images use for experiment.

Makefile
========
When SNAKE_STANDALONE is defined in snake.h
`make snake` makes the `snake`: core algorithm.
When SNAKE_STANDALONE is not defined in snake.h
`make` makes the `main`: main experiment.